[RFR-15] Add Storage to Google Cloud (#154)
* Cleanup Detekt reports

This commit should cleanup all the problems reported by Detekt static code analysis and leave a clean repository to continue work on.

* Move clean up operation to its own class

This makes it easier to provide the correct parameters, since they will be different between cloud and grid fs implementations.

* Refactor Config class

Move many of the options from the API project to the collector, to modify them locally.
Also refactored the Config class to a Kotlin implementation.

* Add Authenticator from API project

To make it possible to use the new Config class, this file has been moved from the API project to here.

* Add PRECONDITION_REQUIRED status code for Authenticator

* Use NIO idiomatic way to check for upload folder existence

* Add DataStorageService for Google Cloud Storage

* Fix static code analysis warnings

* Add Google Client dependency as BOM with most recent version

* Add missing import to GridFSStorageIT

* Create and cleanup temporary folder for GridFSStorageIT

* Add inline documentation

* Use new Vertx Credentials class

This makes the authentication code much more type safe, as no JsonObject is passed arround anymore.

* Remove unecessary dependency

This dependency is there from a prior try to create the API directly from its specification, which did not work. We might retry it some time with the most recent version of Vertx, but until then this dependency is not necessary.

* Add logging to application startup

This makes it easier to actually see what happens during application startup.

* Remove uncommented code

Some unecessary old code was committed by accident. Removing it here.

* Improve test case for resumable uploads

This actually checks the events returned during resumable uploads and checks whether those events contain the correct data.

* Separate request creation from execution in login test

This makes the test easier to debug

* Suppress AnnotationSpacing warnings as we can not fix that

* Update Vertx Version in README.md

* Add correct header to Authenticator.kt