[RFR-874] Update Build Environment (#164)
* Rename .java to .kt

* Update Build Process

Makes everything using Kotlin with the most recent dependencies

* Update Docker Setup to use Eclipse Temurin 17 JRE

* Update Github Build Process to use Temurin JDK 17

* Fix Detekt style warnings.

* Remove two deprecation warnings for using the old CompositeFuture

* Add toolchain repository to make the build Gradle 8.0 compatible.

* Remove unused import in GridFsStorageServiceBuilder

* Remove custom mongo driver version

This is not necessary anymore, since Vert.x fixed the Bug we reported in Version 4.5.0

* Add remaining Detekt problems to baseline

These errors are not fixable at the moment. So we add them to the baseline, which means they are ignored by Detekt from now on.

* Enable execution via ./gradlew run

* Fix errors in entrypoint

* Add fixed logback file for docker dev setup

* Add specific Exception if OAuth2 Discovery failed

* Remove embedded String from MainVerticle

* Fix CI build process

* Change Application back to Java.

This does not work on Kotlin as a Jar file. Therefore we changed it back to pure Java for now.

* Ignore GoogleCloudStorageTest temporarily

* Add Keycloak ios theme to development keycloak

* Remove extra from build.gradle.kts

This is not really necessary for a single project setup and only blows up the file.

* Remove commented code from build.gradle.kts

* Revert ignoring GoogleCloudStorageTest