Fix Upload to Google Cloud
Upload used to be not complete, due to ReadStream closing as soon as everything was read. However it is necessary to wait for the WriteStream to also having finished. Thus I introduced a listener to the write stream and configured the input pipe to only close after the WriteStream has processed everything.